requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
alembic==1.15.1
pydantic-settings==2.8.1
asyncpg==0.30.0
python-dotenv==1.0.1
PyJWT==2.10.1
cryptography==45.0.2
httpx==0.28.1
slowapi==0.1.9
redis[asyncio]==5.2.1
apscheduler==3.10.4
